Implemented literal, length/distance frequency tables.
In order to create huffman trees we need to collect statistics on the
frequency of each literal, <distance,length> pointer. The deflate
specification requires 2 different huffman trees, one for length &
literals and one for distances.

Next: Use the technique in the moffat paper to create the huffman codes
which sort the frequencies and then creates the tree in single pass.