Completed acceleration table decoding
* There was a persistent bug that I introduced by doing the order of
  bits consumed incorrectly. I was consuming additional bits everytime I
  was reading a subtable entry instead of peeking those extra varbits
  and then consuming the right amount of bits based on the returned
  entry.
* Revamped the decode function so that it automatically handles
  subtable links.

Note: the more I work on this, the more I realize that you can just
structure this like a finite-state machine where the main loop is just a
giant switch statement depending on the different flags that are set.
There are only so many states: precode values, literals, lengths,
distances, etc.