Redo decompress with new htree.zig
This is a complete rewrite of the decompress implementation. Basically,
I took inspiration to rewrite the decode_table huffman tree so that we
create 3 distinct huffman tree types:

1. precode tree
2. literals & length tree
3. distance tree

This allows for each htree struct to be a little better suited to each
of the 3 htrees since all 3 are needed for each decoding block.

Another thing is that I changed the main decode_table build loop from
using a series of while loops to first go through symbols that would
be in the main decode table, and then a separate one to go through the
extended subtables to using a single `for` loop to go through each
statement.

Basically, these changes were inspired by the new deflate implementation
by ianic in the zig std. I just kept the use of a reversed code so that
a lookup is a single bitwise `&` instead of having to @bitReverse the
input. I also retained the use of subtables so that we can't need to
keep a linked list of the codes. This saves on memory and improves
lookup times