Fix bug in codeword generation
Codewords are incremented by one for codewords of the same length.
However, once the codeword length increases, then the codeword is
shifted over by one. The previous implementation had a bug where the
codeword wouldn't shift over by the right amount if there was a gap in
the codewords, where a codeword of a certain length (x) had a frequency
of zero, but the codeword of length: x+1 had a nonzero frequency.