Change the way the codeword is implemented
Previous, the codeword was right-justified. But I found from the new
deflate implementation in zig std that you can left justify the
codeword.

When a codeword is left justified, the codes look like the following:

0000 0000
0000 0010
0000 0011
0000 0100 ...

When the code is right justified, it looks like:
0000 0000
0010 0000
0011 0000
0100 0000

This means that to increment the codeword, all we need to do is to do
`codeword += 1 << (maxlen - codeword_length)`

An advantage to this is that this just works even when there are gaps
where a codeword length has no codes but the codeword length above it
has codes. In the first implementation, you need to shift the codewords
to the left twice, but not for the new implementation.

This saves on having to keep track of the current codeword length for
each codeword that we're on