reimplementation from scratch day 2
Today, I'm able to create the basic lookup table for doing very simple
lookup decodes. For the example, this seems to work well enough.
However, I need to create a larger test set to really test cases when
the lookup table isn't big enough to hold all of the codes.

Other notable changes is the use of `Codeword` as a packed struct.
However, I may need to add fields for when the decoder needs information
about whether the lookup table has the right symbol, or whether it needs
to traverse the linked list to get to the right symbol, because its code
doesn't fit in the lookup.