htree from scratch day 4
I realize now that I need to have the symbol table basically be sorted
by the lengths. This is so that the next pointers can work. Also, in
general there are other benefits for sorting the symbols by length.

Right now, I'm getting the offsets table incorrect. I know I need to use
the frequency table, but next time I do this, I should get a pen and
paper and really think about it because I'm not getting it right in the
first try.