htree from scratch day 5
One of the things that I realized yesterday is that with this method of
where I am storing the indexes of the codes as reversed into the lookup
table, I kind of lose the ability to use next pointers. So this was a
very basic attempt at to just make the code tables, but this time in
sorted order with the offset table.

I need to take another look at the lookup subtables for handling codes
with lengths that would mean that it wouldn't fit in the main lookup
table.