htree2 implement next linked list for non-fast lookups
This wasn't a rewrite from scratch, but merely me inplementing some
parts that I needed to have a previous implementation for. But now that
I know it works, I need to do a rewrite.

The key discovery is that I can keep links to the symbols table from the
fast lookup table.

For example:

symbol D has a code (.len=10) that cannot fit in our lookup table for fast
decoding (can store up to 8 bits, that is codes with a length of up to
8). So we take the prefix of symbol D (the first 8 bits) and store a
flag so that we know that this needs a level of indirection. Then we
store an index to the symbol tables.

Since we may have > 1 code with that exact prefix, it would be best for
us to also use a linked list method of prepending entries in the list.

```
var p = &lookup[prefix];
const temp = p.next;
p.len = 0;
p.next = symbol_idx;
symbol[symbol_idx].next = temp;
```

One thing that I've already thought about is that this linked list is
prepending symbols in ascending order of code lengths. This means that
during the decoding process, going off the happy path of the fast lookup
table means traversing the linked list starting with the code with the
most bits, which is the opposite of what we want. Huffman codes are
based on symbol frequency statistics and so it is therefore more likely
that the code we are looking for has fewer bits. So it would be better
to build the lookup table from the end of the symbols table to the
beginning.