Fix bug related to query
Initially, my query function was working because I was testing codes
that are the same reversed. But since I wasn't storing the reversed
codes in the symbol table, the comparison between the bits in the buffer
and the codeword were failing.

For my next attempt, I need to smooth out some of the implementation in
query and try to improve the elegence of the query function.