Implement increment with left lsb for reverse codes
Instead of having to create the codes with lsb on the right, we can
increment the codes with lsb on the left. Ultimately though, reversing
the bits is a little more readable and is only done with the creation of
the symbols table which happens once per decoding section.

It was a fun exercise though.