htree2.zig rewrite day 7
This is a complete rewrite that generates the codes with lsb on the
left. We can do this because we first sort all the codes canonically
into the codeword table and then we can scan through in an easy loop.

This reduces the need to do a @bitReverse on the codes after deriving
them the normal way. This diverges from the example that's in the RFC.
The linked list for decoding symbols that cannot fit in the lookup table
remains the same.