Bugfixes
Biggest issue was that I wasn't advancing the string index on a string
match with length.

Incorrect invariant for the next pointer. Next could be 0, so the
inequality was not correct.

Bug with possible collision on substring match length measuring. This
creates a (0,0) pair.

Added in handling possible chars at the end of string buffer.