lz77 rewrite day 3
* Added an arraylist for appending tokens (literals or (dist,len) pairs).

Retrospect

I'm wondering what a good design for handling the termination of chain
pointers outside the valid range. Usually array indices are `usize` but
when required to do various kinds of math, it's beneficial to use
something like an `i32`.

With a signed integer, a `-1` basically acts as a high bit flag. We
could use an enum for this

```
const chain_i = enum(u16) {
    empty = std.math.maxInt(u16),
    _,

    fn is_empty(self: chain_i) bool {
        return (self & self.empty) == self.empty;
    }

    fn to_int(self: chain_i) usize {
        return @intFromEnum(self);
    }
}
```

Would this make the code more unwieldy? It's something to explore.

AMENDMENT #1
After exploring this option, it doesn't seem to help with making the
code that much. However, perhaps there is an invariant that I'm missing
from my current use of an i32.

AMENDMENT #2
Since the windows for something like DEFLATE is 32kb, which is 15 bits
of information, using an i16 would be possible because the top bit would
be for checking the loop invariant for list pointers, while the bottom
15 bits are able to index into enough slots for the chain list.

There's a certain amount of elegance in making sure to use the right
integer types so that your loop invariants work, but also that other
kinds of invariants such as bitshifts and array indexing also work.