lz77 rewrite day 4
Not a complete rewrite but fundamentally changed the way that
backreference pointers are followed.

Previously, the `chain` table would contain a `usize` to the absolute
index of a u8 in a string. This is expensive; 8 bytes is too much.
Instead we can do much better by storing the i16 of the `index mod
window`.

Each pointer essentially points back a relative offset of the current
index. But a complication emerges because the chain table is also a
circular buffer. Termination isn't just for when you follow a pointer to
an element with a negative value, but if you are further away from the
cursor index (`i`) than the set search window (which in this case is
2^15 elements).

Also fix an issue where `next` pointers weren't correctly read from the
right table.