lz77 rewrite day 5: improve search loop invariant elegance
The key insight for the search loop invariant is that the distance
between our cursor head and a back reference must not exceed the size of
our window.

This reduces a null/negative check in the chain table which also reduces
the need for using a signed integer. Instead, the a chain entry can
point to itself which will terminate the loop because the distance
between chain[i] and i will be the search window size (see fn distance
which calculates a circular buffer distance between two indices in a
single direction with wrapping).

This also helps simplify the types because I no longer have to use
signed integers and can remove potential bugs of underflow.